add untested non-git gemm convolution: compiles
almost identical to jit stuff, except:
- no packed structs needed in any_ vs jit_primitive.hpp
- jit --> any in class names (but many are identical (FIXME))
- jit sgemm dynamic pointer-to-function
  becomes plain old call to SGEMM (fortran interface) as in external/
      mkl_blas.h
- FIXME XXX untested XXX FIXME

+++ build.sh
@@ -171,11 +171,11 @@ else
     echo "Build NOT OK..."
 fi
 echo "DOVANILLA=${DOVANILLA}, DOJIT=${DOJIT}, DOTEST=${DOTEST}, DODEBUG=${DODEBUG}, DODOC=${DODOC}, DONEEDMKL=${DONEEDMKL}"
-if [ "$DOTEST" -gt 0 ]; then
+if [ "$DOTEST" -gt 0 -a "${BUILDOK}" == "y" ]; then
     LOGDIR="log-${DOVANILLA}${DOJIT}${DOTEST}${DODEBUG}${DODOC}${DONEEDMKL}"
     echo "LOGDIR:       ${LOGDIR}" 2>&1 >> build.log
 fi
-if [ "$DOTEST" -gt 0 ]; then
+if [ "$DOTEST" -gt 0  -a "${BUILDOK}" == "y"]; then
     if [ -d "${LOGDIR}" ]; then rm -f "${LOGIDR}.bak"; mv -v "${LOGDIR}" "${LOGDIR}.bak"; fi
     mkdir ${LOGDIR}
     for f in build.log test1.log test2.log doxygen.log; do
